reorganize spawnables and comment
