--- thetawave_lib/src/spawnable/blast/components/mod.rs
@@ -1,3 +1,0 @@
-mod blast;
-
-pub use self::blast::BlastComponent;
