--- thetawave_lib/src/spawnable/blast/mod.rs
@@ -1,1 +1,0 @@
-pub mod components;
