+++ thetawave_lib/src/spawnable/components/mob.rs
@@ -1,0 +1,22 @@
+use crate::{entities::SpawnableType, resources::DropRolls};
+use amethyst::ecs::prelude::{Component, DenseVecStorage};
+use serde::{Deserialize, Serialize};
+
+/// Used for data unique to mob entities
+#[derive(Clone, Serialize, Deserialize, Debug)]
+pub struct MobComponent {
+    /// Damage to defense upon crossing bottom arena border
+    pub defense_damage: f32,
+    /// Damage dealt to colliding entities
+    pub collision_damage: f32,
+    /// Allied with player or not
+    pub allied: bool, // TODO: remove
+    /// Possible drops
+    pub drop_rolls: DropRolls,
+    // Type of mob
+    pub spawnable_type: SpawnableType, // TODO: change to MobType
+}
+
+impl Component for MobComponent {
+    type Storage = DenseVecStorage<Self>;
+}
