+++ thetawave_lib/src/spawnable/components/mod.rs
@@ -1,0 +1,10 @@
+//! Components unique to spawnable entities
+
+mod blast;
+mod consumable;
+mod item;
+mod mob;
+
+pub use self::{
+    blast::BlastComponent, consumable::ConsumableComponent, item::ItemComponent, mob::MobComponent,
+};
