--- thetawave_lib/src/spawnable/consumable/components/mod.rs
@@ -1,3 +1,0 @@
-mod consumable;
-
-pub use self::consumable::ConsumableComponent;
