--- thetawave_lib/src/spawnable/consumable/mod.rs
@@ -1,1 +1,0 @@
-pub mod components;
