--- thetawave_lib/src/spawnable/item/components/mod.rs
@@ -1,3 +1,0 @@
-mod item;
-
-pub use self::item::ItemComponent;
