--- thetawave_lib/src/spawnable/item/mod.rs
@@ -1,1 +1,0 @@
-pub mod components;
