--- thetawave_lib/src/spawnable/mob/components/mod.rs
@@ -1,3 +1,0 @@
-mod mob;
-
-pub use self::mob::MobComponent;
