--- thetawave_lib/src/spawnable/mob/mod.rs
@@ -1,2 +1,0 @@
-pub mod components;
-pub mod systems;
