--- thetawave_lib/src/spawnable/mob/systems/mob_behavior.rs
@@ -1,34 +1,0 @@
-use crate::{
-    components::HealthComponent, events::MobDestroyedEvent,
-    spawnable::mob::components::MobComponent,
-};
-use amethyst::{
-    ecs::prelude::{Entities, Join, System, Write, WriteStorage},
-    shrev::EventChannel,
-};
-
-pub struct MobBehaviorSystem;
-
-impl<'s> System<'s> for MobBehaviorSystem {
-    type SystemData = (
-        Entities<'s>,
-        WriteStorage<'s, MobComponent>,
-        WriteStorage<'s, HealthComponent>,
-        Write<'s, EventChannel<MobDestroyedEvent>>,
-    );
-
-    fn run(
-        &mut self,
-        (entities, mut mobs, mut healths, mut mob_destroyed_event_channel): Self::SystemData,
-    ) {
-        for (mob_entity, _mob_component, mob_health) in (&*entities, &mut mobs, &mut healths).join()
-        {
-            mob_health.constrain();
-
-            // conditions for despawning
-            if mob_health.value <= 0.0 {
-                mob_destroyed_event_channel.single_write(MobDestroyedEvent::new(mob_entity));
-            }
-        }
-    }
-}
