--- thetawave_lib/src/spawnable/mob/systems/mod.rs
@@ -1,4 +1,0 @@
-mod mob_behavior;
-mod mob_destroyed;
-
-pub use self::{mob_behavior::MobBehaviorSystem, mob_destroyed::MobDestroyedSystem};
