+++ thetawave_lib/src/spawnable/mod.rs
@@ -1,7 +1,4 @@
-pub mod blast;
-pub mod consumable;
+pub mod components;
 pub mod entities;
-pub mod item;
-pub mod mob;
 pub mod resources;
 pub mod systems;
