+++ thetawave_lib/src/spawnable/resources/mod.rs
@@ -1,7 +1,8 @@
+//! Resources for spawnable entities
+
 mod spawnables;
 
 pub use self::spawnables::{
-    ConsumableEntityData, ConsumablesResource, EffectEntityData, EffectsResource, GibletEntityData,
-    ItemEntityData, ItemsResource, MobEntityData, MobsResource, RandomMotionRange2D,
-    ThrusterEntityData,
+    ConsumableEntityData, ConsumablesResource, EffectEntityData, EffectsResource, ItemEntityData,
+    ItemsResource, MobEntityData, MobsResource, RandomMotionRange2D, ThrusterEntityData,
 };
