+++ thetawave_lib/src/spawnable/systems/mod.rs
@@ -1,2 +1,7 @@
+mod mob;
 mod modifiers;
-pub use self::modifiers::ModifiersSystem;
+
+pub use self::{
+    mob::{MobBehaviorSystem, MobDestroyedSystem},
+    modifiers::ModifiersSystem,
+};
